// src/app/pages/admin/components/users/components/usuarios-table/user.interface.ts
export interface User {
    nickname: string;
    usuarioauth: string;
    claveauth: string;
    idempresa: number;
    idrol: number;
    usuario: string;
    contrasena: string;
    nombre: string;
    email: string;
    telefono: string;
    costo: number;
    idstatususuario: number;
    emailsms: string;
    bfechainicial: boolean;
    fechainicial: string;
}